### 2016-10-01 #77 @MarkEdmondson1234
Pre-application enquiry - googleComputeEngineR
Hello, I'm in the process of writing an interface with Google Cloud Compute, and have been looking at the docker code from https://github.com/sckott/analogsea that led me to here.

As `analogsea` isn't included in `ropensci` perhaps that answers my question already, but I have a lot of docker stuff in/planned https://github.com/cloudyr/googleComputeEngineR that I felt would fall under the reproducibility criteria of rOpenSci.

For example, it has VM templates that pull from a docker container using [cloudconfig](https://github.com/cloudyr/googleComputeEngineR/blob/master/inst/cloudconfig/rstudio.yaml) to create the initial state.  These config files can be set up to pull your own docker image on boot.  I want to implement it so those docker containers are saved to your private/public google project with your code + data in its frozen state, which can be fired up with a bigger machine on GCE or pulled from another application running docker.

Longer term I'd also like to implement Kubernetes clusters, which allow reproducibility for clusters of docker images, and Dataproc, which is a Spark cluster API.

My main motivation for applying is to get some serious review of the code and help to prioritise features.

Thanks for reading!
Mark
