### 2016-10-03 #77 @sckott
hi @MarkEdmondson1234 - It looks like you're asking if the docker stuff would fit in here? If so: probably. One of our own (Carl Boettiger) co-maintains Rocker https://github.com/rocker-org - so it's stuff we care about

@wch did start a package consolidating docker stuff in https://github.com/wch/harbor - though not on CRAN.  Maybe it'd be worth collaborating with him on it.  I imagine `harbor` would want to stick to just docker stuff, so Kubernetes would be a sep. package.
