### 2016-10-03 #77 @MarkEdmondson1234
Thanks @sckott , yes I have had a look through `harbor` and think I will depend on it if it can be CRANd or at least copy functions.

Great though, I'll keep it in mind when finishing up for a first CRAN release that would suit the criteria.  One thought though is that its now part of cloudyr project - do those clash at all?  Perhaps a fork could live in each?
