### 2016-10-03 #77 @sckott
@MarkEdmondson1234 Sorry, I meant any code dealing with docker itself, not google compute engine, is what's of interest here.
